# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Societal Innovation Portal API
  version: 1.0.0
  description: >
    Canonical machine-readable contract for the SIH 2026 Societal Innovation Portal.
    Human-facing notes, examples and socket events live in docs/API_CONTRACT.md.
    Request/response shapes are generated from packages/shared-types (Zod) --
    if the two disagree, packages/shared-types wins and this file is the bug.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://report-maro-c8m1.onrender.com
    description: Deployed demo
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/proposals:
    get:
      tags: [Proposals]
      summary: List proposals
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/ProposalStatus' } }
        - { in: query, name: assignmentId, schema: { type: string } }
        - { in: query, name: universityId, schema: { type: string } }
        - { in: query, name: mine, schema: { type: boolean } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Paginated proposals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Proposal' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
    post:
      tags: [Proposals]
      summary: Submit a proposal for an accepted assignment (UNIVERSITY)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProposalRequest' }
      responses:
        '201':
          description: Proposal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Proposal' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/proposals/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Proposals]
      summary: Proposal detail
      responses:
        '200':
          description: Proposal
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Proposal' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Proposals]
      summary: Edit a DRAFT or submit it (UNIVERSITY, owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                summary: { type: string }
                approach: { type: string }
                teamMembers:
                  type: array
                  items: { $ref: '#/components/schemas/TeamMember' }
                budgetInr: { type: integer }
                timelineWeeks: { type: integer }
                status: { type: string, enum: [DRAFT, SUBMITTED] }
      responses:
        '200':
          description: Updated proposal
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Proposal' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/proposals/{id}/review:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      tags: [Proposals]
      summary: Approve or reject a proposal (ADMIN)
      description: Approving creates a Project in AWAITING_FUNDING and notifies industries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [APPROVED, REJECTED] }
                reviewNote: { type: string, maxLength: 1000 }
      responses:
        '200':
          description: Reviewed proposal (projectId set when approved)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Proposal' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/projects/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Projects]
      summary: Project detail with funding ledger
      security: []
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalKey:
      type: apiKey
      in: header
      name: x-internal-key
  responses:
...
    Forbidden:
      description: Authenticated but the role is not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Illegal state transition or duplicate
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [success, error, requestId]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, PAYLOAD_TOO_LARGE, INTERNAL_ERROR, SERVICE_UNAVAILABLE]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }
        requestId: { type: string }
    ListMeta:
      type: object
      properties:
        pagination:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    ProblemCategory:
      type: string
      enum: [WATER_SANITATION, ROADS_TRANSPORT, EDUCATION, HEALTHCARE, ELECTRICITY, WASTE_MANAGEMENT, AGRICULTURE, PUBLIC_SAFETY, OTHER]
    Priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    ProblemStatus:
      type: string
      enum: [SUBMITTED, PROCESSING, TRIAGED, DUPLICATE, ASSIGNED, IN_PROGRESS, RESOLVED, REJECTED]
...
    ProposalStatus: { type: string, enum: [DRAFT, SUBMITTED, APPROVED, REJECTED] }
    ProjectStatus: { type: string, enum: [AWAITING_FUNDING, ACTIVE, COMPLETED, CANCELLED] }
    OrganizationSummary:
      type: object
      required: [id, name, type, district]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [UNIVERSITY, INDUSTRY, GOVERNMENT] }
        district: { type: [string, 'null'] }
...
    Location:
      type: object
      required: [latitude, longitude, district]
      properties:
        latitude: { type: number, minimum: 21.5, maximum: 25.6 }
        longitude: { type: number, minimum: 83.2, maximum: 88.0 }
        district: { type: string }
        address: { type: string }
...
    ProblemSummary:
      type: object
      required: [id, title, status, location, reporter, category, priority, duplicateCount, assignmentId, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/ProblemStatus' }
        location: { $ref: '#/components/schemas/Location' }
        reporter:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        category:
          oneOf:
            - $ref: '#/components/schemas/ProblemCategory'
            - type: 'null'
        priority:
          oneOf:
            - $ref: '#/components/schemas/Priority'
            - type: 'null'
        duplicateCount: { type: integer }
        assignmentId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TeamMember:
      type: object
      required: [name, role]
      properties:
        name: { type: string }
        role: { type: string }
        email: { type: string, format: email }
    Proposal:
      type: object
      required: [id, assignmentId, problem, university, title, summary, approach, teamMembers, budgetInr, timelineWeeks, status, reviewNote, reviewedAt, projectId, submittedBy, createdAt, updatedAt]
      properties:
        id: { type: string }
        assignmentId: { type: string }
        problem: { $ref: '#/components/schemas/ProblemSummary' }
        university: { $ref: '#/components/schemas/OrganizationSummary' }
        title: { type: string }
        summary: { type: string }
        approach: { type: string }
        teamMembers: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
        budgetInr: { type: integer }
        timelineWeeks: { type: integer }
        status: { $ref: '#/components/schemas/ProposalStatus' }
        reviewNote: { type: [string, 'null'] }
        reviewedAt: { type: [string, 'null'], format: date-time }
        projectId: { type: [string, 'null'] }
        submittedBy:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateProposalRequest:
      type: object
      required: [assignmentId, title, summary, approach, teamMembers, budgetInr, timelineWeeks]
      properties:
        assignmentId: { type: string }
        title: { type: string, minLength: 10 }
        summary: { type: string, minLength: 50 }
        approach: { type: string, minLength: 50 }
        teamMembers: { type: array, minItems: 1, items: { $ref: '#/components/schemas/TeamMember' } }
        budgetInr: { type: integer, minimum: 0 }
        timelineWeeks: { type: integer, minimum: 1 }
        status: { type: string, enum: [DRAFT, SUBMITTED], default: SUBMITTED }
    Funding:
      type: object
      required: [id, projectId, industry, amountInr, note, fundedBy, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        industry: { $ref: '#/components/schemas/OrganizationSummary' }
        amountInr: { type: integer }
        note: { type: [string, 'null'] }
        fundedBy:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        createdAt: { type: string, format: date-time }
    ProjectSummary:
      type: object
      required: [id, proposalId, problem, university, title, category, status, budgetInr, fundedInr, fundedPercent, timelineWeeks, startedAt, completedAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        proposalId: { type: string }
        problem: { $ref: '#/components/schemas/ProblemSummary' }
        university: { $ref: '#/components/schemas/OrganizationSummary' }
        title: { type: string }
        category:
          oneOf:
            - $ref: '#/components/schemas/ProblemCategory'
            - type: 'null'
        status: { $ref: '#/components/schemas/ProjectStatus' }
        budgetInr: { type: integer }
        fundedInr: { type: integer }
        fundedPercent: { type: number }
        timelineWeeks: { type: integer }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          required: [summary, fundings]
          properties:
            summary: { type: string }
            fundings: { type: array, items: { $ref: '#/components/schemas/Funding' } }
...